+++ src/main/java/wooteco/subway/section/service/SectionService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.exception.RequestException;
-import wooteco.subway.section.controller.dto.SectionCreateDto;
-import wooteco.subway.section.controller.dto.SectionDeleteDto;
-import wooteco.subway.section.controller.dto.SectionDto;
+import wooteco.subway.section.service.dto.SectionCreateDto;
+import wooteco.subway.section.service.dto.SectionDeleteDto;
+import wooteco.subway.section.service.dto.SectionDto;
 import wooteco.subway.section.domain.Section;
 import wooteco.subway.section.domain.SectionRepository;
 import wooteco.subway.section.domain.Sections;
